*refactoring of ConvertLogSoftmaxNode.
Move ConvertLogSoftmaxNode from OnnxRuntime.h into ConvertLogSoftmaxNode.cpp
and register the new source in the CvtOnnx VS2022 project.
