+++ src/Cvt/OnnxRuntime/Convert.h
@@ -115,6 +115,8 @@ namespace Synet
 
     bool ConvertLogNode(const onnx::NodeProto& node, LayerParam& layer);
 
+    bool ConvertLogSoftmaxNode(const onnx::NodeProto& node, bool trans, const LayerParams& layers, const Bytes& original, LayerParam& layer);
+
     bool ConvertMulNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, const OnnxParam& onnxParam, LayerParam& layer);
 
     bool ConvertNonMaxSuppressionNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& bin, LayerParam& layer);
